Use existing cmake file where present (#64)
* check for existing cmake file in specified test dir

* fix linting warning in entrypoint script

* target dockerfile instead of image

* unpin dependency versions

* Update test_released_action.yml

* adjust entrypoint script

* re-pin dependency versions

* add unit test project with existing cmake

* update cmake version

* only set cmake policy CMP0135 when using cmake ver after 3.23

* debug message if test exes could not be found

* rename dummy test project

* reintroduce test action for project without existing cmake file

* add coverage compiler option

* try setting gcovr root dir to source

* revert gcovr root dir

* add source dir argument to action

* fix entrypoint script

* target pre-built image

* increment pre release version

* increment pre release version

* switch to full release image reference

* prune example CMakeLists

* test updated CMakeLists

* revert to using release image

* remove unneeded cmake command

* add test case for mixed project style

* force image build on run for testing

* revert to release image

* update readme

---------

+++ entrypoint.sh
@@ -9,16 +9,38 @@ IFS=';' read -ra paths <<< "$1"
 for i in "${paths[@]}"; do
   echo "processing unit tests in: $i"
   
-  cp /CMakeLists.txt "$i"
+  # Check if CMakeLists.txt already exists in the unit test directory
+  if [ ! -e "$i/CMakeLists.txt" ]; then
+    cp /CMakeLists.txt "$i"
+  fi
+
   pushd "$i" || exit 1
   cmake CMakeLists.txt
   make || exit 2
 
-  # execute the unit tests
+  # Find all executables ending in _tests recursively in the specified root directory
   popd || exit 3
-  valgrind --leak-check=full --error-exitcode=4 "./$i/unit_tests" || exit 4
+  mapfile -t test_executables < <(find "$i" -type f -executable -name "*_tests")
+
+  # Check if any test executables were found
+  if [ ${#test_executables[@]} -eq 0 ]; then
+    echo "no test executables found"
+    exit 4
+  fi
+
+  # execute the unit tests
+  for executable in "${test_executables[@]}"; do
+    valgrind --leak-check=full --error-exitcode=4 "$executable" || exit 4
+  done
+
+  # determine source dir
+  if [ -z "$2" ]; then
+    source_dir="."
+  else
+    source_dir=$2
+  fi
 
   # generate a coverage report
   uuid=$(uuidgen)
-  gcovr --branches --xml-pretty -r . >> "$uuid-report.xml"
+  gcovr . --root "$source_dir"  --branches --xml-pretty >> "$uuid-report.xml"
 done
